Fix Postgres parameter type errors in leave approval and fleet assignment.
Cast SQL parameters explicitly for enum and integer fields to avoid 42P08 inconsistent parameter type failures on Render when approving leave or assigning vehicles.

Made-with: Cursor

+++ lib/fleet.ts
@@ -57,5 +57,8 @@ export async function createVehicle(input:{type:string;beschrijving:string|null;
 }
 
 export async function assignVehicle(vehicleId:number,userId:number|null) {
-  await pool.query(`UPDATE apparatuur SET toegewezen_aan=$2,toewijzingsdatum=CASE WHEN $2 IS NULL THEN NULL ELSE CURRENT_DATE END WHERE id=$1`, [vehicleId,userId]);
+  await pool.query(
+    `UPDATE apparatuur SET toegewezen_aan=$2::int,toewijzingsdatum=CASE WHEN $2::int IS NULL THEN NULL ELSE CURRENT_DATE END WHERE id=$1`,
+    [vehicleId,userId],
+  );
 }
